requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.4
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap4==22.1
django-crispy-forms==1.14.0
idna==3.3
lxml==4.9.0
multitasking==0.0.10
numpy==1.22.4
pandas==1.4.2
Pillow==9.1.1
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.9
Wand==0.6.
scipy==1.10.1
